+++ OGDotNet.Tests.Integration/ValueAssertions.cs
@@ -336,7 +336,7 @@ namespace OGDotNet.Tests.Integration
         {
             Assert.NotNull(position);
             Assert.NotNull(position.UniqueId);
-            Assert.InRange(position.Quantity, -20, long.MaxValue); //Not sure why -20 is valid (Bond Future Test View)
+            Assert.InRange(position.Quantity, -20, decimal.MaxValue); //Not sure why -20 is valid (Bond Future Test View)
             AssertSensibleValue(position.SecurityKey);
             Assert.NotNull(position.Trades);
             AssertSensibleValue(position.Trades);
